Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ int main(void)
 	long long int mylli;
 	float myfloat;
 
-	printf("Size of char: %lu byte(s)\n", sizeof(mychar));
-	printf("Size of int: %lu byte(s)\n", sizeof(myint));
-	printf("Size of long int: %lu byte(s)\n", sizeof(myli));
-	printf("Size of long long int: %lu byte(s)\n", sizeof(mylli));
-	printf("Size of float: %lu byte(s)\n", sizeof(myfloat));
+	printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(mychar));
+	printf("Size of int: %lu byte(s)\n", (unsigned long)sizeof(myint));
+	printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(myli));
+	printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(mylli));
+	printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(myfloat));
 
 	return (0);
 }
